Reorganized imports
I reorganized the imports so the module is more self-contained.
Ideally, we would get rid of the led_controller directory and move
everything up a level so that __init__.py and plugin.info appear
in the root directory. That way when you git-clone it, the
plugin.info file is right inside the directory git creates which
is just named for your project.

We generally don't build new plugins in the Naomi/plugins directory
any more. The Naomi Plugin Exchange (NPE) puts plugins into the
~/.configs/naomi/plugins/<plugin_type> directory so that is where
I build new ones. That makes handling the plugin as an independent
git repository easier, instead of working with submodules in Naomi.

I got rid of the absolute import and replaced it with a few
relative imports, to make this module self-contained.

I'm working on a virtual machine, so once I put in a request it
just generates an error saying the RPC is unavailable. Please let
me know if it works.